Moved common client into own package.
Refactored client implementation to strategy pattern.

Updated package to only expose public Factory function.
