+++ clients/client/local.go
@@ -1,0 +1,34 @@
+/*******************************************************************************
+ * Copyright 2019 Dell Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ *******************************************************************************/
+
+package client
+
+import (
+	"github.com/edgexfoundry/go-mod-core-contracts/clients/types"
+)
+
+type Client struct {
+	url string
+}
+
+// newLocal returns a pointer to a Client.
+func newLocal(params types.EndpointParams) *Client {
+	return &Client{
+		url: params.Url,
+	}
+}
+
+func (c *Client) URL() (string, error) {
+	return c.url, nil
+}
